Do not TRIM() parentheses around partial indexe conditions
Because PostgreSQL will return the expression with a lot of
parentheses we cannot TRIM() them easily. It is easier and more
correct to adapt to what PostgreSQL returns. That means the declaration
of partial indexes must be updated as follow:

Before:
``options={"where": "other_id IS NULL"}``

After:
``options={"where": "(other_id IS NULL)"}``

And fore more complexe conditions, that would be:
``options={"where": "(((id IS NOT NULL) AND (other_id IS NULL)) AND (name IS NULL))"}``